Move md2 devices to mxbluesky.devices.md2
