+++ start_bs.py
@@ -8,7 +8,7 @@ from mxtools.eiger import EigerSingleTriggerV26, set_eiger_defaults
 import os
 from mxtools.governor import _make_governors
 from ophyd.signal import EpicsSignalBase
-from devices import LightDevice, BeamstopDevice, MD2SimpleHVDevice, MD2Device, ShutterDevice
+from mxbluesky.devices.md2 import LightDevice, BeamstopDevice, MD2SimpleHVDevice, MD2Device, ShutterDevice
 EpicsSignalBase.set_defaults(timeout=10, connection_timeout=10)  # new style
 from mxbluesky.devices import (WorkPositions, TwoClickLowMag, LoopDetector, MountPositions, 
                                  TopAlignerFast, TopAlignerSlow, GoniometerStack)
